+++ R/kruskal_wallis.R
@@ -1,6 +1,6 @@
 #' Hettmansperger-Norton Trend Test for k-Samples
 #'
-#' @description This function calculates the Hettmansperger-Norton trend test using pseudo-ranks under the null hypothesis H0F: F_1 = ... F_k = 0.
+#' @description This function calculates the Kruskal-Wallis test using pseudo-ranks under the null hypothesis H0F: F_1 = ... F_k.
 #' @param data numeric vector containing the data
 #' @param group factor specifying the groups
 #' @param na.rm a logical value indicating if NA values should be removed
@@ -9,7 +9,7 @@
 #' @param ... further arguments are ignored
 #' @return Returns a data.frame with the results
 #' @example R/example_3.txt
-#' @references Brunner, E., Bathke A. C. and Konietschke, F. Rank- and Pseudo-Rank Procedures in Factorial Designs - Using R and SAS. Springer Verlag. to appear.
+#' @references Brunner, E., Bathke, A.C., and Konietschke, F. (2018a). Rank- and Pseudo-Rank Procedures for Independent Observations in Factorial Designs - Using R and SAS. Springer Series in Statistics, Springer, Heidelberg. ISBN: 978-3-030-02912-8.
 #' @references Hettmansperger, T. P., & Norton, R. M. (1987). Tests for patterned alternatives in k-sample problems. Journal of the American Statistical Association, 82(397), 292-299
 #' @keywords internal
 kruskal_wallis_internal <- function(data, group, na.rm, formula = NULL, pseudoranks = TRUE, ...) {
@@ -31,10 +31,10 @@ kruskal_wallis_internal <- function(data, group, na.rm, formula = NULL, pseudora
   N <- sum(n)
   a <- length(n)
   if(pseudoranks){
-    df <- data.frame(pranks = psrank(data, group), group = group)
+    df <- data.frame(pranks = pseudorank(data, group), group = group)
     
   } else {
-    df <- data.frame(pranks = psrank(data, as.factor(rep(1, length(data)))), group = group)
+    df <- data.frame(pranks = pseudorank(data, as.factor(rep(1, length(data)))), group = group)
   }
   df <- df[order(df$group),]
   R_mean <- summaryBy(pranks~group,data=df, FUN = mean)[, 2]
@@ -50,6 +50,8 @@ kruskal_wallis_internal <- function(data, group, na.rm, formula = NULL, pseudora
   output <- list()
   output$name <- "Kruskal-Wallis Test"
   output$test <- Q_N
+  output$distribution <- "Chisq"
+  output$df <- a-1
   output$pValue <- pValue
   output$ss <- n
   output$pHat <- 1/N*(R_mean-1/2)