+++ NEWS.md
@@ -1,3 +1,7 @@
+# pseudorank 0.3.7
+* Changed function name ''psrank'' to ''pseudorank'' and set ''psrank'' as deprecated.
+* Added a data set which can be used to show paradoxical results from rank tests.
+
 # pseudorank 0.3.0
 
 * Added the Kruskal-Wallis test with an option to use either ranks or pseudo-ranks.